project/requirements.txt
opencv-python==4.11.0.86
numpy==2.2.2
Pillow==11.1.0
matplotlib==3.10.0
PyYAML==6.0.2
